+++ llama_index/bridge/langchain.py
@@ -26,7 +26,7 @@ from langchain.memory import ChatMessageHistory, ConversationBufferMemory
 
 # chat and memory
 from langchain.memory.chat_memory import BaseChatMemory
-from langchain.output_parsers import PydanticOutputParser, ResponseSchema
+from langchain.output_parsers import ResponseSchema
 from langchain.prompts.chat import (
     AIMessagePromptTemplate,
     BaseMessagePromptTemplate,
@@ -90,7 +90,6 @@ __all__ = [
     "Tool",
     "BaseTool",
     "ResponseSchema",
-    "PydanticOutputParser",
     "print_text",
     "get_color_mapping",
     "BaseCallbackHandler",