+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### New Features
+- Add CorrectnessEvaluator (#7661)
 - Added support for `Ollama` LLMs (#7635)
 - Added `HWPReader` (#7672)
 - Simplified portkey LLM interface (#7669)
@@ -14,6 +15,15 @@
 - Imrpoved reliability of LLM selectors (#7678)
 - Fixed `query_wrapper_prompt` and `system_prompt` for output parsers and completion models (#7678)
 - Fixed node attribute inheritance in citation query engine (#7675)
+
+### Breaking Changes
+- Refactor and update `BaseEvaluator` interface to be more consistent (#7661)
+  - Use `evaluate` function for generic input
+  - Use `evaluate_response` function with `Response` objects from llama index query engine
+- Update existing evaluators with more explicit naming
+  - `ResponseEvaluator` -> `FaithfulnessEvaluator`
+  - `QueryResponseEvaluator` -> `RelevancyEvaluator`
+  - old names are kept as class aliases for backwards compatibility
 
 ## [0.8.27] - 2023-09-14